Libft and gnl

+++ src/Libft/ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: gmarchal <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 14:52:50 by gmarchal          #+#    #+#             */
-/*   Updated: 2022/10/10 15:34:39 by gmarchal         ###   ########.fr       */
+/*   Updated: 2023/04/24 12:33:00 by gmarchal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, int len)
 {
-	size_t			i;
+	int			i;
 	unsigned char	val;
 	unsigned char	*ret;
