+++ src/Libft/ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: gmarchal <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/06 11:19:42 by gmarchal          #+#    #+#             */
-/*   Updated: 2022/10/06 15:27:25 by gmarchal         ###   ########.fr       */
+/*   Updated: 2023/04/24 12:31:25 by gmarchal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, int len)
 {
-	size_t	i;
-	size_t	j;
+	int		i;
+	int		j;
 	char	*ret;
 
 	i = -1;
